Create header using flexbox
Done using space-between which puts left and right link at edges
but logo is spaced perfectly in the middle between the two.
Padding is also provided to the links arent exactly on the edge of
the container. Got rid of margin and padding in ul tags. Separated
li tags using gap.