Create choice selection for computer and user
The choice selection for computer and user are two functions named
getComputerChoice() and getHumanChoice(). getComputerChoice() uses
Math.floor() and Math.random() to randomly choose a number 0-2
randomly, and that number is passed to the gameChoices() function
which will return its choice. For the getHumanChoice() function,
the user is prompted with a menu and asks to select their choice
via the number corresponding to the choice. After a number is
selected (1-3 are the options) it is converted to a number with
parseInt() and decremented by one so it can be passed to
gameChoices(). It is decremented by 1 because the indexes for the
choices in the menu are off by 1.