Create game logic and result checker
resultCheck() function checks the results from the choices
both the human and computer chose and returns whether the
human won, lost, or tied with the computer. This result is
used the the playRound() function which tells you if you
won, lost or tied the round and returns the result from
resultCheck(). playGame() function keeps track of score and
calls to get the choices. It also calls the playRound()
function 5 times from a for loop to make it a game of 5.
In the end, the playGame() function displays the final
score out of 5 games.