Solve exercise 4 from javascript exercises
Learned how to use filter, and includes method along
with arrow functions. arrow function is commonly used
with filter method to take specific values out. include
was used to see if the value to be removed was in the
array passed to the function. the ! (not operator) was
used to negate the falsy values to truthy because we
only want to pass values that ARENT in the remove
argument. with .includes we are looking for the values
in the remove argument.